refactor(TKT-329): consolidate repo/repos namespace to singular repo namespace with --bulk flag
