+++ apps/cli/src/commands/repo/add.ts
@@ -4,8 +4,10 @@ import { colors, format } from '../../lib/colors.js';
 import {
   findHQRoot,
   promptAddSingleRepo,
+  promptForRepositories,
   addRepository
 } from '../../lib/repos/index.js';
+import { getWorkspaceRepositories } from '../../lib/database/index.js';
 
 export default class Add extends PMOCommand {
   static description = 'Add a repository to the HQ';
@@ -14,6 +16,7 @@ export default class Add extends PMOCommand {
     '<%= config.bin %> <%= command.id %> /path/to/repo',
     '<%= config.bin %> <%= command.id %> git@github.com:user/repo.git',
     '<%= config.bin %> <%= command.id %>',
+    '<%= config.bin %> <%= command.id %> --bulk',
   ];
 
   static args = {
@@ -31,6 +34,11 @@ export default class Add extends PMOCommand {
       options: ['clone', 'move'],
       default: 'clone',
     }),
+    bulk: Flags.boolean({
+      char: 'b',
+      description: 'Add multiple repositories interactively',
+      default: false,
+    }),
   };
 
   protected getPMOOptions() {
@@ -44,6 +52,12 @@ export default class Add extends PMOCommand {
     const hqPath = findHQRoot();
     if (!hqPath) {
       this.error('Not in an HQ directory. Run "prlt init" first.');
+    }
+
+    // Bulk mode: add multiple repositories interactively
+    if (flags.bulk) {
+      await this.executeBulk(hqPath);
+      return;
     }
 
     let repoPath: string;
@@ -78,6 +92,51 @@ export default class Add extends PMOCommand {
       this.log(format.success(`Repository ${result.name} added successfully`));
     } else {
       this.error(`Failed to add repository: ${result.error}`);
+    }
+  }
+
+  /**
+   * Bulk mode: add multiple repositories interactively
+   */
+  private async executeBulk(hqPath: string): Promise<void> {
+    this.log(colors.primary('📦 Add Repositories (Bulk Mode)\n'));
+
+    // Get existing repos
+    const existingRepos = getWorkspaceRepositories(hqPath).map(r => r.name);
+
+    // Use the shared prompt for repositories
+    const reposToAdd = await promptForRepositories(process.cwd(), existingRepos);
+
+    if (reposToAdd.length === 0) {
+      this.log(colors.textMuted('No repositories selected.'));
+      return;
+    }
+
+    this.log('');
+
+    // Add each repository
+    let successCount = 0;
+    let failCount = 0;
+
+    for (const repo of reposToAdd) {
+      const result = await addRepository(hqPath, repo.path, repo.action);
+
+      if (result.success) {
+        this.log(format.success(`Repository ${result.name} added`));
+        successCount++;
+      } else {
+        this.log(format.error(`Failed to add ${result.name}: ${result.error}`));
+        failCount++;
+      }
+    }
+
+    // Summary
+    this.log('');
+    if (successCount > 0) {
+      this.log(format.success(`Added ${successCount} repository(ies)`));
+    }
+    if (failCount > 0) {
+      this.log(format.error(`Failed to add ${failCount} repository(ies)`));
     }
   }
 }
